[fix] rescue Net::IMAP::NoResponseError in webmail mailboxes_controller (#2876)

+++ app/models/webmail/mailbox.rb
@@ -26,9 +26,6 @@ class Webmail::Mailbox
   validates :name, presence: true
 
   before_validation :validate_name, if: ->{ @sync && name_changed? }
-  before_create :imap_create, if: ->{ @sync && imap.present? }
-  before_update :imap_update, if: ->{ @sync && imap.present? }
-  before_destroy :imap_delete, if: ->{ @sync && imap.present? }
 
   default_scope -> { order_by order: 1, downcase_name: 1 }
 
@@ -116,24 +113,15 @@ class Webmail::Mailbox
     self.depth = original_name.split(box.delim).size - 1
   end
 
-  private
-
-  def validate_name
-    self.name = self.name.tr('/', '.')
-    #self.name = "INBOX.#{name}" unless self.name =~ /^INBOX\./
-    self.original_name = Net::IMAP.encode_utf7(name)
-    self.delim = '.'
-    self.attr = []
-    self.depth = self.name.split('.').size - 1
-  end
-
   def imap_create
+    return false if imap.blank?
     imap.conn.create original_name
   rescue Net::IMAP::NoResponseError => e
     rescue_imap_error(e)
   end
 
   def imap_update
+    return false if imap.blank?
     imap.conn.rename original_name_was, original_name
     items = Webmail::Mail.where(imap.account_scope).where(mailbox: name_was)
     items.each(&:destroy_rfc822)
@@ -143,12 +131,24 @@ class Webmail::Mailbox
   end
 
   def imap_delete
+    return false if imap.blank?
     imap.conn.delete original_name
     items = mails
     items.each(&:destroy_rfc822)
     items.delete_all
   rescue Net::IMAP::NoResponseError => e
     rescue_imap_error(e)
+  end
+
+  private
+
+  def validate_name
+    self.name = self.name.tr('/', '.')
+    #self.name = "INBOX.#{name}" unless self.name =~ /^INBOX\./
+    self.original_name = Net::IMAP.encode_utf7(name)
+    self.delim = '.'
+    self.attr = []
+    self.depth = self.name.split('.').size - 1
   end
 
   def rescue_imap_error(exception)
